+++ moneta/src/python/helper/apps.py
@@ -1,0 +1,5 @@
+from django.apps import AppConfig
+
+
+class ExpendAppConfig(AppConfig):
+    name = 'helper'
