+++ moneta/src/python/helper/pool_manager.py
@@ -5,7 +5,6 @@ import threading
 import functools
 import MySQLdb
 
-
 LAST_UPDATE = 'last_update'
 CREATE_TIME = 'create_time'
 CONNECTION = 'connection'
@@ -143,5 +142,3 @@ def pool_manage():
             if DB_POOL[DATA_BASE] is None:
                 DB_POOL[DATA_BASE] = DBPoolManager(**CONNECT_SETTINGS)
     return DB_POOL[DATA_BASE]
-
-
