+++ moneta/src/python/moneta/settings.py
@@ -55,6 +55,7 @@ INSTALLED_APPS = [
     'income',
     'expend',
     'helper',
+    'tests'
 ]
 
 MIDDLEWARE = [
