+++ moneta/src/python/tests/apps.py
@@ -1,0 +1,5 @@
+from django.apps import AppConfig
+
+
+class ExpendAppConfig(AppConfig):
+    name = 'tests'
