+++ moneta/src/python/tests/test_redis_worker.py
@@ -3,7 +3,7 @@ import unittest
 
 import nose
 
-from redis_worker import RedisWorker
+from src.python.helper.redis_worker import RedisWorker
 
 
 class TestRedisWorker(unittest.TestCase):
